refactor(settings): place agent actions in backend
